extract tenant id calculation to allow unit testing
