+++ test/extensions/tracers/opentelemetry/samplers/dynatrace/BUILD
@@ -32,6 +32,7 @@ envoy_extension_cc_test(
         "sampler_config_test.cc",
         "sampling_controller_test.cc",
         "stream_summary_test.cc",
+        "tenant_id_test.cc",
     ],
     extension_names = ["envoy.tracers.opentelemetry.samplers.dynatrace"],
     deps = [
